Refactor remote config into module
