+++ Iquiz-assets/main.js
@@ -1,6 +1,7 @@
 import { $, $$ } from './src/utils/dom.js';
 import { clamp, faNum, faDecimal, formatDuration, formatRelativeTime } from './src/utils/format.js';
 import { configureFeedback, vibrate, toast, wait, SFX, shootConfetti } from './src/utils/feedback.js';
+import { RemoteConfig, patchPricingKeys } from './src/config/remote-config.js';
 
   // Anti-cheating: Detect devtools
   (function() {
@@ -61,98 +62,6 @@ function populateProvinceOptions(selectEl, placeholder){
 
   
   const online = () => navigator.onLine;
-  
-  // ===== Remote Config (in-file) =====
-// === RemoteConfig (اصلاح‌شده) ===
-const RemoteConfig = {
-  ab: (Math.random() < 0.5) ? 'A' : 'B',
-
-  provinceTargeting: { enabled: true, allow: ['تهران','کردستان','آذربایجان غربی','اصفهان'] },
-
-  ads: {
-    enabled: true,
-    placements: { banner:true, native:true, interstitial:true, rewarded:true },
-    freqCaps: { interstitialPerSession: 2, rewardedPerSession: 3 },
-    interstitialCooldownMs: 60_000,
-    rewardedMinWatchMs: 7_000,
-    session: { interstitialShown: 0, rewardedShown: 0, lastInterstitialAt: 0 }
-  },
-
-  pricing: {
-    // نرخ تبدیل دلاری برای مواقعی که priceToman نداشتیم (اختیاری)
-    usdToToman: 70_000,
-
-    // ← این بخش به تومان به‌روزرسانی شد
-    coins: [
-      { id:'c100',  amount:100,  bonus:0,  priceToman: 59_000,   priceCents:199  },
-      { id:'c500',  amount:500,  bonus:5,  priceToman: 239_000,  priceCents:799  },
-      { id:'c1200', amount:1200, bonus:12, priceToman: 459_000,  priceCents:1499 },
-      { id:'c3000', amount:3000, bonus:25, priceToman: 899_000,  priceCents:2999 }
-    ],
-
-    vip: {
-      lite: { id:'vip_lite', priceCents:299 },
-      pro:  { id:'vip_pro',  priceCents:599 }
-    },
-
-    // بسته‌های «کلید» (پرداخت با سکهٔ بازی)
-    keys: [
-      { id:'k1',  amount:1,  priceGame:30  }, // بسته کوچک
-      { id:'k3',  amount:3,  priceGame:80  }, // بسته اقتصادی
-      { id:'k10', amount:10, priceGame:250 }  // بسته بزرگ
-    ]
-  },
-
-  abOverrides: {
-    A: { ads:{ freqCaps:{ interstitialPerSession:2 } } },
-    B: { ads:{ freqCaps:{ interstitialPerSession:1 } } }
-  },
-
-  gameLimits: {
-    matches: { daily: 5, vipMultiplier: 2, recoveryTime: 2 * 60 * 60 * 1000 }, // 2 hours
-    duels:   { daily: 3, vipMultiplier: 2, recoveryTime: 30 * 60 * 1000 },     // 30 minutes
-    lives:   { daily: 3, vipMultiplier: 2, recoveryTime: 30 * 60 * 1000 },     // 30 minutes
-    groupBattles: { daily: 2, vipMultiplier: 2, recoveryTime: 60 * 60 * 1000 }, // 1 hour
-    energy:  { daily: 10, vipMultiplier: 2, recoveryTime: 15 * 60 * 1000 }     // 15 minutes
-  }
-};
-
-
-
-// === Patch ایمن برای RemoteConfig.pricing.keys ===
-// اگر keys نباشه/خالی یا نامعتبر باشه، با دیفالت‌ها پر و نرمالایز می‌کنه.
-// خروجی: آرایه‌ی keys نهایی (مرتب و سالم)
-function patchPricingKeys(RemoteConfig){
-  const defaults = [
-    { id:'k1',  amount:1,  priceGame:30,  label:'بسته کوچک'     },
-    { id:'k3',  amount:3,  priceGame:80,  label:'بسته اقتصادی'   },
-    { id:'k10', amount:10, priceGame:250, label:'بسته بزرگ'      }
-  ];
-
-  if (!RemoteConfig.pricing) RemoteConfig.pricing = {};
-
-  const packs = RemoteConfig.pricing.keys;
-  const bad = (p)=> typeof p?.amount!=='number' || typeof p?.priceGame!=='number' || p.amount<=0 || p.priceGame<=0;
-
-  if (!Array.isArray(packs) || packs.length===0 || packs.some(bad)){
-    RemoteConfig.pricing.keys = defaults;
-  } else {
-    RemoteConfig.pricing.keys = packs
-      .map(p => ({
-        id: String(p.id || ('k' + p.amount)),
-        amount: +p.amount,
-        priceGame: +p.priceGame,
-        label: p.label || (p.amount<=1 ? 'بسته کوچک' : p.amount<=3 ? 'بسته اقتصادی' : 'بسته بزرگ')
-      }))
-      .filter(p => p.amount>0 && p.priceGame>0)
-      .sort((a,b)=> a.amount - b.amount);
-  }
-
-  return RemoteConfig.pricing.keys;
-}
-
-// ⚡️ بلافاصله بعد از تعریف RemoteConfig صدا بزن:
-patchPricingKeys(RemoteConfig);
 
   const DEFAULT_DIFFS = [
     { value: 'easy', label: 'آسان' },
@@ -529,14 +438,6 @@ patchPricingKeys(RemoteConfig);
     }catch{}
   }
 
-  // Merge AB overrides
-  (function applyAB(){
-    const o = RemoteConfig.abOverrides[RemoteConfig.ab];
-    if(!o) return;
-    function deepMerge(t,s){ for(const k in s){ if(typeof s[k]==='object'&&s[k]){ t[k]=t[k]||{}; deepMerge(t[k],s[k]); } else { t[k]=s[k]; } } }
-    deepMerge(RemoteConfig, o);
-  })();
-  
   // ===== Analytics =====
   async function logEvent(name, payload={}){
     try{
